LayoutTests/storage/websql/alter-to-info-table.html

<html>
<head>
<script src="resources/database-common.js"></script>
<script src="alter-to-info-table.js"></script>
</head>
<body onload="setupAndRunTest();">
This tests that a tricky way to setup your own info table fails.<br>
<pre id="console">
FAILURE: test didn't run.
</pre>
</body>
</html>
