LayoutTests/storage/websql/alter-to-info-table-expected.txt

CONSOLE MESSAGE: line 8: Web SQL is deprecated. Please use IndexedDB instead.
This tests that a tricky way to setup your own info table fails.
CREATE TABLE statement succeeded.
CREATE TEMP TABLE statement succeeded.
INSERT IN TEMP TABLE statement succeeded.
CREATE TRIGGER statement succeeded.
ALTER TO INFO TABLE statement succeeded.
Step 1 transaction succeeded.
Successfully changed DB version
Step 2 transaction succeeded.
